EFCore,MySQL,Connection updates
As I'm unable to fix connectivity issues between FiveM and MSSQL, changed the database provider to MySql. Still researching on how to do it with MSSQL.

Updated EF6 to EFCore.
Created a resource for validating the user using playerConnect event, provided by FiveM.
I've only implemented gathering the licenses. Next step is to check for bans.

WARNING: Everything is on development state. Actually I'm still testing the .NET in FXServer, so it will takes some time until the whole framework is ready!